[INLONG-5299][Manager] Remove redundant configuration and class files
