+++ inlong-manager/manager-service/src/test/java/org/apache/inlong/manager/pojo/common/utils/ValidationUtilsTest.java
@@ -18,7 +18,7 @@
 package org.apache.inlong.manager.pojo.common.utils;
 
 import org.apache.inlong.manager.pojo.cluster.ClusterTagRequest;
-import org.apache.inlong.manager.pojo.common.UpdateValidation;
+import org.apache.inlong.manager.common.validation.UpdateValidation;
 import org.apache.inlong.manager.common.util.ValidationUtils;
 import org.apache.inlong.manager.service.ServiceBaseTest;
 import org.junit.jupiter.api.Assertions;
