+++ inlong-manager/manager-test/src/main/resources/application-unit-test.properties
@@ -30,13 +30,6 @@ spring.datasource.url=jdbc:h2:mem:test;MODE=MYSQL;DB_CLOSE_DELAY=-1;IGNORECASE=T
 spring.datasource.username=root
 spring.datasource.password=""
 
-# Manager address of Tube cluster, used to create Topic
-cluster.tube.manager=http://127.0.0.1:8081
-# Master address, used to manage Tube broker
-cluster.tube.master=127.0.0.1:8000,127.0.0.1:8010
-# Tube cluster ID
-cluster.tube.clusterId=1
-
 # Audit configuration
 # Audit query source that decide what data source to query, currently only supports [MYSQL|ELASTICSEARCH]
 audit.query.source=MYSQL
