+++ inlong-manager/manager-web/src/main/resources/application-test.properties
@@ -48,13 +48,6 @@ spring.datasource.druid.connectionProperties=druid.stat.mergeSql=true;druid.stat
 # Exclude ElasticsearchRestClientAutoConfiguration
 spring.autoconfigure.exclude=org.springframework.boot.autoconfigure.elasticsearch.ElasticsearchRestClientAutoConfiguration
 
-# Manager address of Tube cluster, used to create Topic
-cluster.tube.manager=http://127.0.0.1:8081
-# Master address, used to manage Tube broker
-cluster.tube.master=127.0.0.1:8000,127.0.0.1:8010
-# Tube cluster ID
-cluster.tube.clusterId=1
-
 # Audit configuration
 # Audit query source that decide what data source to query, currently only supports [MYSQL|ELASTICSEARCH]
 audit.query.source=MYSQL
