+++ inlong-manager/manager-web/src/test/resources/application.properties
@@ -25,7 +25,6 @@ server.servlet.context-path=/inlong/manager
 spring.application.name=InLong-Manager-Web
 spring.profiles.active=dev
 
-spring.main.allow-circular-references=true
 spring.mvc.pathmatch.matching-strategy=ANT_PATH_MATCHER
 
 # Serialize the Date type to a timestamp
